requirements.txt
google-generativeai>=0.5.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.90.0
gspread>=5.10.0
requests>=2.31.0
python-dateutil>=2.8.0
